% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{GetGenesData}
\alias{GetGenesData}
\title{GetGenesData}
\usage{
GetGenesData(gff, vcf)
}
\description{
Metapoly's import function, takes as input a gff file open via read.delim (see
example on github), and a VCF file loaded with vcfR (read.vcfR). Returns the
polymorphism data in the format used for all MetaPoly's analyses.
}
