% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_corr.R
\name{PolyCorr}
\alias{PolyCorr}
\title{PolyCorr}
\usage{
PolyCorr(data, min_samp, samp_vec, p_corr = "holm")
}
\description{
This functions computes the gene-wise correlation between the number of SNPs, and a target
variable than can be either continuous, either discrete (handles two treatments). It creates
a zero-inflated poisson model of the SNP count, taking the sample, log(sequencing depth),
and the gene length as fixed effects. The idea is to identify genes showing signs of selective
sweeps associated with the target variable. There are three inputs:
\itemize{
\item data: the polymorphism data created by MetaPoly's "GetGenesData" function
\item min_samp: the minimal number of sample having the required depth to run the correlation test.
\item samp_vec: a vector that assigns each sample (vector values) to each population (vector
names). The method handles two populations encoded as 0 and 1's.
}
}
