% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_div.R
\name{PolyDiv}
\alias{PolyDiv}
\title{PolyDiv}
\usage{
PolyDiv(data, samp_vec)
}
\description{
This functions computes Fst at the gene level, between two sets of metagenomic samples. It
compares the average allele frequencies across samples for each polymorphic site, and then
outputs the median Fst value. The inputs are:
\itemize{
\item data: the polymorphism data created by MetaPoly's "GetGenesData" function
\item samp_vec: a vector that assigns each sample (vector values) to each population (vector
names). The method handles two populations encoded as 0 and 1's
}
}
