% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{PolySummary}
\alias{PolySummary}
\title{PolySummary}
\usage{
PolySummary(data, samp_vec)
}
\description{
Summarises the polymorphism in the different genes, across samples. The number
of SNP (SNP_N), sequencing depth (DEPTH), major allele frequency (MAJF), and
allele frequencies evenness (EVENNESS) are measured and returned in a large df.
}
