% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{SummariseSamples}
\alias{SummariseSamples}
\title{SummariseSamples}
\usage{
SummariseSamples(poly_summary)
}
\description{
Summarises the polymorphism data generated by PolySummary by sample, to get genome-
wide data per sample
}
